+++ lib/brando/blueprint/json_ld/json_ld_field.ex
@@ -1,0 +1,5 @@
+defmodule Brando.Blueprint.JSONLD.JSONLDField do
+  defstruct name: nil,
+            type: nil,
+            value_fn: nil
+end