+++ lib/brando/blueprint/json_ld/json_ld_schema.ex
@@ -1,0 +1,5 @@
+defmodule Brando.Blueprint.JSONLD.JSONLDSchema do
+  defstruct __identifier__: nil,
+            schema: nil,
+            fields: []
+end