+++ lib/brando/plugs/html.ex
@@ -80,10 +80,15 @@ defmodule Brando.Plug.HTML do
     do: assign(conn, :json_ld_breadcrumbs, breadcrumbs)
 
   def put_json_ld(conn, module, data, extra_fields \\ []) do
-    extra_fields = JSONLD.convert_format(extra_fields)
-    meta_meta = %{__meta__: %{current_url: Utils.current_url(conn)}}
+    meta_meta = %{
+      __meta__: %{
+        current_url: Utils.current_url(conn),
+        language: Map.get(data, :language, conn.assigns.language)
+      }
+    }
+
     data_with_meta = Map.merge(data, meta_meta)
-    assign(conn, :json_ld_entity, module.extract_json_ld(data_with_meta, extra_fields))
+    assign(conn, :json_ld_entity, JSONLD.extract_json_ld(module, data_with_meta, extra_fields))
   end
 
   @doc """