+++ test/brando/meta/schema_test.exs
@@ -51,20 +51,20 @@ defmodule Brando.MetaSchemaTest do
 
   test "fallback" do
     data = %{meta_title: "META title", title: "Title", foo: "bar"}
-    assert Brando.Blueprint.Meta.fallback(data, [:meta_title, :title]) == "META title"
-    assert Brando.Blueprint.Meta.fallback(data, [:title, :meta_title]) == "Title"
-    assert Brando.Blueprint.Meta.fallback(data, [:title, :foo]) == "Title"
-    assert Brando.Blueprint.Meta.fallback(data, [:foo, :title]) == "bar"
+    assert Brando.Blueprint.Utils.fallback(data, [:meta_title, :title]) == "META title"
+    assert Brando.Blueprint.Utils.fallback(data, [:title, :meta_title]) == "Title"
+    assert Brando.Blueprint.Utils.fallback(data, [:title, :foo]) == "Title"
+    assert Brando.Blueprint.Utils.fallback(data, [:foo, :title]) == "bar"
 
     data = %{meta_title: nil, title: "Title", foo: "bar"}
-    assert Brando.Blueprint.Meta.fallback(data, [:meta_title, :title]) == "Title"
+    assert Brando.Blueprint.Utils.fallback(data, [:meta_title, :title]) == "Title"
 
     data = %{meta_title: nil, title: nil, foo: "bar"}
-    assert Brando.Blueprint.Meta.fallback(data, [:meta_title, :title]) == nil
+    assert Brando.Blueprint.Utils.fallback(data, [:meta_title, :title]) == nil
 
     data = %{title: "<p>Title grabbed from a field with rich text</p>", foo: "bar"}
 
-    assert Brando.Blueprint.Meta.fallback(data, [:meta_title, {:strip_tags, :title}]) ==
+    assert Brando.Blueprint.Utils.fallback(data, [:meta_title, {:strip_tags, :title}]) ==
              "Title grabbed from a field with rich text"
   end
 end